pyproject.toml
[project]
name = "pic-format-converter"
version = "0.1.0"
description = "Convert HEIC/HEIF images to JPG format"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=10.0.0",
    "pillow-heif>=0.18.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.scripts]
heic2jpg = "main:app"

[tool.uv]
package = true
